fix: index.ts has worker.js & types are separate
