+++ test/file-generator.test.ts
@@ -1,12 +1,13 @@
 import fs from 'fs';
 
-const directory = './test/tmp';
-const fileCount = 500;
+const directory = './test';
+const fileCount = 100;
 
 describe('File Creation Test', () => {
     beforeAll(() => {
         createTestFiles(directory, fileCount);
     });
+
     afterAll(() => {
         deleteTestFiles(directory, fileCount);
     });
